NSFPAN-1065 Convert 'HttpClientConfig' to Kotlin.

+++ Modules/RoxieMobile.NetworkingApi/networking-api/src/main/java/com/roxiemobile/networkingapi/network/NetworkConfig.java
@@ -7,7 +7,7 @@ public interface NetworkConfig
     interface Timeout
     {
         // In milliseconds
-        int CONNECTION = 60 * 1000;
-        int READ = 30 * 1000;
+        long CONNECTION = 60 * 1000;
+        long READ = 30 * 1000;
     }
 }
